# Improve cost of BLS key serialization for gRPC