@hexfusion 2023-04-19
> lgtm, there may be other places in the codebase where we can use this optimization (matter for a different PR)

Yeah I was going to break out the logic into funcs my fear is someone using the it in the wrong place (like me :)) to grab a perf gain and opening vuln.